added FwFM
