+++ torchfm/test-datasets/dummy.txt
@@ -1,0 +1,20 @@
+1,1,3,5,7,9,11
+1,2,3,5,7,9,11
+0,2,4,6,8,10,12
+0,2,4,5,7,10,11
+1,1,3,5,7,9,11
+1,2,3,5,7,9,11
+0,2,4,6,8,10,12
+0,2,4,5,7,10,11
+1,1,3,5,7,9,11
+1,2,3,5,7,9,11
+0,2,4,6,8,10,12
+0,2,4,5,7,10,11
+1,1,3,5,7,9,11
+1,2,3,5,7,9,11
+0,2,4,6,8,10,12
+0,2,4,5,7,10,11
+1,1,3,5,7,9,11
+1,2,3,5,7,9,11
+0,2,4,6,8,10,12
+0,2,4,5,7,10,11
