+++ torchfm/torch_utils/utils.py
@@ -1,0 +1,128 @@
+import torch
+
+from torchfm.dataset.avazu import AvazuDataset
+from torchfm.dataset.criteo import CriteoDataset
+from torchfm.dataset.movielens import MovieLens1MDataset, MovieLens20MDataset
+from torchfm.dataset.dummy_dataset import DummyDataset
+from torchfm.model.afi import AutomaticFeatureInteractionModel
+from torchfm.model.afm import AttentionalFactorizationMachineModel
+from torchfm.model.dcn import DeepCrossNetworkModel
+from torchfm.model.dfm import DeepFactorizationMachineModel
+from torchfm.model.ffm import FieldAwareFactorizationMachineModel
+from torchfm.model.fm import FactorizationMachineModel
+from torchfm.model.fnfm import FieldAwareNeuralFactorizationMachineModel
+from torchfm.model.fnn import FactorizationSupportedNeuralNetworkModel
+from torchfm.model.hofm import HighOrderFactorizationMachineModel
+from torchfm.model.lr import LogisticRegressionModel
+from torchfm.model.ncf import NeuralCollaborativeFiltering
+from torchfm.model.nfm import NeuralFactorizationMachineModel
+from torchfm.model.pnn import ProductNeuralNetworkModel
+from torchfm.model.wd import WideAndDeepModel
+from torchfm.model.xdfm import ExtremeDeepFactorizationMachineModel
+from torchfm.model.afn import AdaptiveFactorizationNetwork
+from torchfm.model.fwfm import FieldWeightedFactorizationMachineModel
+
+
+def get_criterion(criterion):
+    if criterion == 'bceloss':
+        return torch.nn.BCELoss()
+    elif criterion == 'bcelogitloss':
+        return torch.nn.BCEWithLogitsLoss()
+    elif criterion == 'mseloss':
+        return torch.nn.MSELoss()
+    elif criterion == "nllloss":
+        return torch.nn.NLLLoss()
+    else:
+        raise ValueError('unknown criterion name: ' + criterion)
+
+
+def get_dataset(name, path):
+    if name == 'movielens1M':
+        return MovieLens1MDataset(path)
+    elif name == 'movielens20M':
+        return MovieLens20MDataset(path)
+    elif name == 'criteo':
+        return CriteoDataset(path)
+    elif name == 'avazu':
+        return AvazuDataset(path)
+    elif name == 'dummy':
+        return DummyDataset(path)
+    else:
+        raise ValueError('unknown dataset name: ' + name)
+
+
+def get_model(name, dataset):
+    """
+    Hyperparameters are empirically determined, not opitmized.
+    """
+    field_dims = dataset.field_dims
+    features, labels = iter(dataset).__next__()
+    num_columns = len(features)
+
+    if name == 'lr':
+        return LogisticRegressionModel(field_dims)
+    elif name == 'fm':
+        return FactorizationMachineModel(field_dims, embed_dim=16)
+    elif name == 'hofm':
+        return HighOrderFactorizationMachineModel(field_dims, order=3, embed_dim=16)
+    elif name == 'ffm':
+        return FieldAwareFactorizationMachineModel(field_dims, embed_dim=4)
+    elif name == 'fwfm':
+        return FieldWeightedFactorizationMachineModel(field_dims, embed_dim=4, num_factors=num_columns)  # num_factors - num entries in  a row
+    elif name == 'fnn':
+        return FactorizationSupportedNeuralNetworkModel(field_dims, embed_dim=16, mlp_dims=(16, 16), dropout=0.2)
+    elif name == 'wd':
+        return WideAndDeepModel(field_dims, embed_dim=16, mlp_dims=(16, 16), dropout=0.2)
+    elif name == 'ipnn':
+        return ProductNeuralNetworkModel(field_dims, embed_dim=16, mlp_dims=(16,), method='inner', dropout=0.2)
+    elif name == 'opnn':
+        return ProductNeuralNetworkModel(field_dims, embed_dim=16, mlp_dims=(16,), method='outer', dropout=0.2)
+    elif name == 'dcn':
+        return DeepCrossNetworkModel(field_dims, embed_dim=16, num_layers=3, mlp_dims=(16, 16), dropout=0.2)
+    elif name == 'nfm':
+        return NeuralFactorizationMachineModel(field_dims, embed_dim=64, mlp_dims=(64,), dropouts=(0.2, 0.2))
+    elif name == 'ncf':
+        # only supports MovieLens dataset because for other datasets user/item colums are indistinguishable
+        assert isinstance(dataset, MovieLens20MDataset) or isinstance(dataset, MovieLens1MDataset)
+        return NeuralCollaborativeFiltering(field_dims, embed_dim=16, mlp_dims=(16, 16), dropout=0.2,
+                                            user_field_idx=dataset.user_field_idx,
+                                            item_field_idx=dataset.item_field_idx)
+    elif name == 'fnfm':
+        return FieldAwareNeuralFactorizationMachineModel(field_dims, embed_dim=4, mlp_dims=(64,), dropouts=(0.2, 0.2))
+    elif name == 'dfm':
+        return DeepFactorizationMachineModel(field_dims, embed_dim=16, mlp_dims=(16, 16), dropout=0.2)
+    elif name == 'xdfm':
+        return ExtremeDeepFactorizationMachineModel(
+            field_dims, embed_dim=16, cross_layer_sizes=(16, 16), split_half=False, mlp_dims=(16, 16), dropout=0.2)
+    elif name == 'afm':
+        return AttentionalFactorizationMachineModel(field_dims, embed_dim=16, attn_size=16, dropouts=(0.2, 0.2))
+    elif name == 'afi':
+        return AutomaticFeatureInteractionModel(
+             field_dims, embed_dim=16, atten_embed_dim=64, num_heads=2, num_layers=3, mlp_dims=(400, 400), dropouts=(0, 0, 0))
+    elif name == 'afn':
+        print("Model:AFN")
+        return AdaptiveFactorizationNetwork(
+            field_dims, embed_dim=16, LNN_dim=1500, mlp_dims=(400, 400, 400), dropouts=(0, 0, 0))
+    else:
+        raise ValueError('unknown model name: ' + name)
+
+
+class EarlyStopper(object):
+
+    def __init__(self, num_trials, save_path):
+        self.num_trials = num_trials
+        self.trial_counter = 0
+        self.best_accuracy = 0
+        self.save_path = save_path
+
+    def is_continuable(self, model, accuracy):
+        if accuracy > self.best_accuracy:
+            self.best_accuracy = accuracy
+            self.trial_counter = 0
+            torch.save(model, self.save_path)
+            return True
+        elif self.trial_counter + 1 < self.num_trials:
+            self.trial_counter += 1
+            return True
+        else:
+            return False
